brought main3 to the same level as WaveCam_SV notebook

+++ main3.py
@@ -1,24 +1,17 @@
+# Importing the relevant packages
 import numpy as np
 import cv2
 import matplotlib.pyplot as plt
 import sys
 import pyaudio
 
-# Courant number
-cn = 0.9 / np.sqrt(2.0)
-# Pi
-pi = 3.14
-# Constants
-alpha = (0.0, 0.0)  # sound attenuation coefficient
-alphap = (0.0, 0.0)  # sound attenuation coefficient
-# Default number of media
-nm = 2
-
 # Ask for user input
+plt.ion()
 dflag = input("Enter 'd' to run the default setup or 'c' for customizing: ")
 print("")
 if dflag == "d":
     freq = 15000
+    nm = 2
     c0 = (346.13, 0)
     rho = (1.2, 1.0e6)
     stype = "point"
@@ -30,41 +23,53 @@ elif dflag == "c":
     sflag = input("Enter 'a' to record an audio source or 'n' to enter a numeric source frequency: ")
     print("")
     if sflag == "a":
-        freq = 0
-        rt = 80000
-        rec_sec = 1
-        chsize = 1024
-        input("Start the audio and then press Enter to record...")
-        print("")
         # Initialize portaudio
         p = pyaudio.PyAudio()
-        astream = p.open(format=pyaudio.paInt16, channels=1, rate=rt, input=True, frames_per_buffer=chsize)
-        # Record audio
-        frames = []
-        for i in range(0, np.int(rt / chsize * rec_sec)):
-            adata = astream.read(chsize)
-            frames.append(np.fromstring(adata, dtype=np.int16))
-        # Close audio stream
-        astream.stop_stream()
-        astream.close()
-        p.terminate()
-        atemp = np.asfarray(np.hstack(frames))
-        skipf = np.alen(atemp) / 4
-        last = np.alen(atemp)
-        asource = ((atemp[skipf:last - skipf] - np.mean(atemp[skipf:last - skipf]))
-                   / (np.amax(atemp[skipf:last - skipf]) - np.amin(atemp[skipf:last - skipf])))
-        sfreqspec = np.fft.rfft(asource)
-        time = np.linspace(0, last, last) * rec_sec / rt
-        plt.figure()
-        plt.plot(time[skipf:last - skipf], asource)
-        plt.xlabel("time (s)")
-        plt.ylabel("rescaled amplitude (arb units)")
-        plt.tight_layout()
-        plt.figure()
-        plt.plot(np.abs(sfreqspec)**2)
-        plt.xlabel("freq (Hz)")
-        plt.ylabel("power (arb units)")
-        plt.tight_layout()
+        numdev = p.get_device_count() - 1
+        if numdev < 1:
+            sys.exit("Error: You do not have the hardware to record audio")
+        else:
+            freq = 0
+            rt = 80000
+            rec_sec = 1
+            chsize = 1024
+            input("Start the audio and then press Enter to record...")
+            print("")
+            astream = p.open(format=pyaudio.paInt16, channels=1, rate=rt, input=True, frames_per_buffer=chsize)
+            # Record audio
+            frames = []
+            for i in range(0, np.int(rt / chsize * rec_sec)):
+                adata = astream.read(chsize)
+                frames.append(np.fromstring(adata, dtype=np.int16))
+            # Close audio stream
+            astream.stop_stream()
+            astream.close()
+            p.terminate()
+            atemp = np.asfarray(np.hstack(frames))
+            skipf = int(np.alen(atemp) / 4)
+            last = int(np.alen(atemp))
+            asource = ((atemp[skipf:last - skipf] - np.mean(atemp[skipf:last - skipf]))
+                       / (np.amax(atemp[skipf:last - skipf]) - np.amin(atemp[skipf:last - skipf])))
+            sfreqspec = np.fft.rfft(asource)
+            time = np.linspace(0, last, last) * rec_sec / rt
+            fig1 = plt.figure()
+            ax1 = fig1.add_subplot(111)
+            ax1.set_xlabel("time (s)")
+            ax1.set_ylabel("rescaled amplitude (arb units)")
+            ax1.set_title("Recorded audio")
+            ax1.plot(time[skipf:last - skipf], asource)
+            fig1.tight_layout()
+            fig1.show()
+            fig1.canvas.draw()
+            fig2 = plt.figure()
+            ax2 = fig2.add_subplot(111)
+            ax2.set_xlabel('freq (Hz)')
+            ax2.set_ylabel('power (arb units)')
+            ax2.set_title("Power spectrum of recorded audio")
+            ax2.plot(np.abs(sfreqspec)**2)
+            fig2.tight_layout()
+            fig2.show()
+            fig2.canvas.draw()
     elif sflag == "n":
         try:
             freq = float(input("Enter the sound frequency in Hz (20-20000): "))
@@ -88,7 +93,7 @@ elif dflag == "c":
             print("")
             rho = (mdensity, 1.0e6)
         except ValueError:
-            sys.exit("Error: enter a numeric value, or 'air', or 'water'")
+            sys.exit("Error: enter a numeric value, or 'air', or 'water' ")
     stype = input("Enter 'point' or 'line' for type of source: ")
     print("")
     if stype != "line" and stype != "point":
@@ -102,12 +107,10 @@ elif dflag == "c":
         ct = c0[0] * np.sqrt(temparature/293)
         c0 = (c0[0], 0, ct)
         rho = (rho[0], 1.0e6, rho[0])
-        alpha = (0.0, 0.0, 0.0)
-        alphap = (0.0, 0.0, 0.0)
     elif mflag == "n":
         nm = 2
     else:
-        sys.exit("Error: enter 'y' to insert a block of different temperature or enter 'n'")
+        sys.exit("Error: enter 'y' to insert a block of different temperature or enter 'n' ")
 else:
     sys.exit("Error: enter 'd' for default setup or 'c' to customize")
 
@@ -117,33 +120,33 @@ else:
     wavelmin = np.amin(c0) / 20000.0
 
 
-class FdtdVar:
+# The main class that defines all constants, variables, and functions
+class fdtdVar:
     def __init__(self, rs, cs):
-        self.r = np.int(rs)
-        self.c = np.int(cs)
-        self.freq = freq
+        # Constants
+        cn = 0.9 / np.sqrt(2.0)  # Courant number
+        # Variables
+        self.r = np.int(rs)  # number of rows
+        self.c = np.int(cs)  # number of columns
+        self.freq = freq  # frequency of source
         temp = (self.r, self.c + 1)
-        self.vx = np.zeros(temp)
+        self.vx = np.zeros(temp)  # velocity along x
         self.mvx = np.zeros(temp, dtype=np.int8)
         temp = (self.r + 1, self.c)
-        self.vy = np.zeros(temp)
+        self.vy = np.zeros(temp)  # velocity along y
         self.mvy = np.zeros(temp, dtype=np.int8)
         temp = (self.r, self.c)
-        self.pr = np.zeros(temp)
+        self.pr = np.zeros(temp)  # pressure
         self.gaussamp = np.zeros(temp)
         self.mpr = np.zeros(temp, dtype=np.int8)
-        self.dx = wavelmin/10.0
-        self.dt = cn * self.dx / np.amax(c0)
+        self.dx = wavelmin / 10.0  # grid cell size
+        self.dt = cn * self.dx / np.amax(c0)  # time step size
         self.ca = np.ones(nm)
         self.cb = np.ones(nm)
         self.da = np.ones(nm)
         self.db = np.ones(nm)
         for i in range(0, nm, 1):
-            # self.ca[i] = ((2 * c0[i]**2 * rho[i] - alpha[i] * self.dt)
-            #               / (2 * c0[i]**2 * rho[i] + alpha[i] * self.dt))
             self.cb[i] = c0[i] ** 2 * rho[i] * self.dt / self.dx
-            # self.da[i] = ((2 * rho[i] - alphap[i] * self.dt)
-            #               / (2 * rho[i] + alphap[i] * self.dt))
             self.db[i] = self.dt / (rho[i] * self.dx)
         self.da[1] = 0
         temp = (self.r, 2, 2)
@@ -179,7 +182,7 @@ class FdtdVar:
             else:
                 prs = 0
         else:
-            prs = self.dx * np.sin(2 * pi * self.freq * nt * self.dt) / self.cb[0]
+            prs = self.dx * np.sin(2 * np.pi * self.freq * nt * self.dt) / self.cb[0]
 
         # Update pressure with source
         self.pr[1:rm - 1, 1:cm - 1] = (self.pr[1:rm - 1, 1:cm - 1]
@@ -189,7 +192,6 @@ class FdtdVar:
     def fdtd_update(self):
         ri = self.r
         ci = self.c
-
         self.pr[0:ri, 0:ci] = (self.ca[self.mpr[0:ri, 0:ci]] * self.pr[0:ri, 0:ci]
                                - self.cb[self.mpr[0:ri, 0:ci]]
                                * ((self.vx[0:ri, 1:ci + 1] - self.vx[0:ri, 0:ci])
@@ -257,8 +259,8 @@ class FdtdVar:
         if mflag == "y":
             cm = self.c
             rm = self.r
-            c1 = np.int(cm/2) + np.int(cm/8)
-            c2 = c1 + np.int(cm/8)
+            c1 = np.int(cm / 2) + np.int(cm / 8)
+            c2 = c1 + np.int(cm / 8)
             self.mvx.fill(0)
             self.mvy.fill(0)
             self.mpr.fill(0)
@@ -271,50 +273,51 @@ class FdtdVar:
             self.mpr.fill(0)
 
 
-def test():
-    t1 = FdtdVar(480, 640)
-    for nt in range(1, 500, 1):
-        t1.fdtd_update()
-        t1.source(nt)
-        t1.boundary()
-        cv2.imshow('frame', t1.pr)
-        # if cv2.waitKey(1) & 0xFF == ord('q'):
-        #     break
-        plt.pcolormesh(t1.pr, cmap="gray_r", vmin=-1, vmax=1)
-        plt.axis("image")
-        plt.colorbar()
-        plt.pause(0.1)
-        plt.clf()
-    plt.show()
+# Ask user for option to read an image from file or start webcam to capture a frame
 
+vflag = input("Do you want to read an image (enter i) or capture a frame from webcam (enter w)? ")
+print("")
+if vflag == "w":
+    # Create VideoCapture object
+    cap = cv2.VideoCapture(0)
 
-# test()
+    # Array sizes (resolution dependent)
+    retval = cap.get(3)
+    columns = retval
+    retval = cap.get(4)
+    rows = retval
 
-# Create VideoCapture object
-cap = cv2.VideoCapture(0)
+    # Set frame resolution
+    if dflag == "c" and columns > 1000:
+        print("Default frame resolution ", np.int(columns), "x", np.int(rows))
+        if input("Do you want to reduce resolution to increase speed? (y/n): ") == "y":
+            print("")
+            print("Setting new resolution to 640 x 480")
+            print("")
+            columns = 640
+            rows = 480
 
-# Array sizes (resolution dependent)
-retval = cap.get(3)
-columns = retval
-retval = cap.get(4)
-rows = retval
-
-# Set frame resolution
-if dflag == "c" and columns > 1000:
-    print('Default frame resolution ', np.int(columns), 'x', np.int(rows))
-    if input("Do you want to reduce resolution to increase speed? (y/n): ") == "y":
-        print("")
-        print('Setting new resolution to 640 x 480')
-        print("")
-        columns = 640
-        rows = 480
-
-cap.set(3, columns)
-cap.set(4, rows)
+    cap.set(3, columns)
+    cap.set(4, rows)
+elif vflag == "i":
+    # Read image
+    imgname = input("Enter the filename of your image including extension: ")
+    print("")
+    img = cv2.imread(imgname, 0)
+    rows, columns = img.shape
+    ar = 640 / columns
+    dim = (640, int(rows * ar))
+    img = cv2.medianBlur(img, 21)
+    resimg = cv2.resize(img, dim, interpolation = cv2.INTER_AREA)
+    rows, columns = resimg.shape
+    print("Resized the image to 640 x 480 pixels")
+    print("")
+    threshimg = cv2.adaptiveThreshold(resimg, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY, 17, 5)
+else:
+    sys.exit("Error: type either 'i' to read image or 'w' to read from webcam")
 
 # Create FDTD object
-fs = FdtdVar(rows, columns)
-
+fs = fdtdVar(rows, columns)
 tc = 0
 
 if dflag == "c":
@@ -326,43 +329,63 @@ else:
     bflag = "c"
 
 # Wait to start wave propagation
-print("Click on the camera window and press 's' to start wave propagation...")
+if vflag == "w":
+    print("Click on the camera window and press 's' to start wave propagation...")
+    print("")
+    while True:
+        # Reset domain
+        fs.update_domain()
+
+        # Capture frame
+        retval, frame = cap.read()
+
+        # Convert to grayscale
+        img = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY) / 256.0
+
+        # Create rigid material
+        imgtemp = np.pad(img, ((0, 0), (0, 1)), "constant", constant_values=1.0)
+        idx = imgtemp < 0.4
+        fs.mvx[idx] = 1
+        imgtemp = np.pad(img, ((0, 1), (0, 0)), "constant", constant_values=1.0)
+        idx = imgtemp < 0.4
+        fs.mvy[idx] = 1
+
+        # Display image
+        cv2.imshow("frame", img)
+        if cv2.waitKey(1) & 0xFF == ord("s"):
+            break
+elif vflag == "i":
+    print("Click on the image window and press 's' to start wave propagation...")
+    print("")
+    while True:
+        # Reset domain
+        fs.update_domain()
+
+        # Capture frame
+        img = threshimg / 256.0
+
+        # Create rigid material
+        imgtemp = np.pad(img, ((0, 0), (0, 1)), "constant", constant_values=1.0)
+        idx = imgtemp < 0.5
+        fs.mvx[idx] = 1
+        imgtemp = np.pad(img, ((0, 1), (0, 0)), "constant", constant_values=1.0)
+        idx = imgtemp < 0.5
+        fs.mvy[idx] = 1
+
+        # Display image
+        cv2.imshow("frame", img)
+        if cv2.waitKey(1) & 0xFF == ord("s"):
+            break
+
+# Release the VideoCapture object
+if vflag == "w":
+    cap.release()
+    cv2.destroyAllWindows()
+
+print("Press 'q' to quit (when the image window is topmost)...")
 print("")
-while True:
-    # Reset domain
-    # fs.mvx.fill(0)
-    # fs.mvy.fill(0)
-    fs.update_domain()
-
-    # Capture frame
-    retval, frame = cap.read()
-
-    # Convert to grayscale
-    img = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY) / 256.0
-
-    # Create rigid material
-    imgtemp = np.pad(img, ((0, 0), (0, 1)), 'constant', constant_values=1.0)
-    idx = imgtemp < 0.4
-    fs.mvx[idx] = 1
-    imgtemp = np.pad(img, ((0, 1), (0, 0)), 'constant', constant_values=1.0)
-    idx = imgtemp < 0.4
-    fs.mvy[idx] = 1
-
-    # Display image
-    cv2.imshow('frame', img)
-    if cv2.waitKey(1) & 0xFF == ord('s'):
-        break
-
-print("Press 'q' to quit (when the camera window is topmost)...")
-print("")
 
 while True:
-    # Capture frame
-    retval, frame = cap.read()
-
-    # Convert to grayscale
-    img = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)/256.0
-
     # Update image with FDTD solution
     fs.fdtd_update()
     fs.source(tc)
@@ -371,16 +394,15 @@ while True:
     imgdisp = img + fs.pr
 
     # Display image
-    cv2.imshow('frame', imgdisp)
-    if cv2.waitKey(1) & 0xFF == ord('q'):
+    cv2.imshow("frame", imgdisp)
+    if cv2.waitKey(1) & 0xFF == ord("q"):
         break
 
     tc = tc + 1
 
-# Release the VideoCapture object
-cap.release()
-cv2.destroyAllWindows()
-
-plt.figure()
-plt.imshow(imgdisp, cmap="gray")
-plt.show()
+fig3 = plt.figure()
+ax3 = fig3.add_subplot(111)
+ax3.imshow(imgdisp, cmap="gray")
+fig3.tight_layout()
+fig3.show()
+fig3.canvas.draw()
